requirements.txt
python-dotenv==1.0.0
Flask==3.1.3
boto3==1.34.0
Werkzeug==3.1.8